Support adding cert to Debian/Ubuntu system store
